Relocate default diagnostic formatting test to a separate file
